refactor(auth): 인증 시스템 개선으로 인한 변경 (#129)
